requirements.txt
Flask
transformers
sentence-transformers
torch
p
pdfkit
Pillow
matplotlib
seaborn
requests
io
os
re
numpy
